+++ NEWS.md
@@ -1,5 +1,11 @@
 # bslib (development version)
 
+## Improvements
+
+* Sidebars now use neutral `<div>` markup by default. Use the new `role`
+  argument of `sidebar()` to identify a sidebar as a complementary landmark,
+  form, search area, or another ARIA role. (#1358)
+
 ## Bug fixes
 
 * Fixed `input_dark_mode()` writing an invalid `"undefined"` value to the `data-bs-theme` attribute when it connected while the document was already set to `"light"`. (#1349)